requirements.txt
appnope==0.1.2
backcall==0.2.0
black==21.12b0
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
coverage==6.2
decorator==5.1.0
Deprecated==1.2.13
fakeredis==1.7.0
idna==3.3
ipdb==0.13.9
ipython==7.30.1
isort==5.10.1
jedi==0.18.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
packaging==21.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.4.0
prompt-toolkit==3.0.24
ptyprocess==0.7.0
Pygments==2.10.0
pyparsing==3.0.6
python-decouple==3.5
redis==4.0.2
requests==2.26.0
six==1.16.0
sortedcontainers==2.4.0
toml==0.10.2
tomli==1.2.3
tornado==6.1
traitlets==5.1.1
typing-extensions==4.0.1
urllib3==1.26.7
wcwidth==0.2.5
wrapt==1.13.3
